+++ src/constants.ts
@@ -16,8 +16,16 @@ export const dependencies = [
  * These dev dependencies are for JavaScript projects.
  */
 export const devDependencies = [
-  "@babel/core",
+  "@types/react",
+  "@types/react-dom",
+  "@types/cors",
+  "@types/express",
+  "@types/mongoose",
+  "@types/morgan",
+  "@types/node-fetch",
   "@babel/cli",
+  "@babel/core",
+  "@babel/plugin-proposal-class-properties",
   "@babel/plugin-transform-runtime",
   "@babel/preset-env",
   "@babel/preset-react",
@@ -37,13 +45,6 @@ export const devDependencies = [
   "webpack",
   "webpack-cli",
   "webpack-dev-server",
-  "@types/react",
-  "@types/react-dom",
-  "@types/cors",
-  "@types/express",
-  "@types/mongoose",
-  "@types/morgan",
-  "@types/node-fetch",
 ];
 
 /**
@@ -74,6 +75,9 @@ export const devDependenciesTS = [
   "webpack-dev-server",
 ];
 
+/**
+ * Files needed for the Build process, to be copied.
+ */
 export const buildFilesToCopy = [
   {
     src: "src/server/assets",
@@ -89,6 +93,9 @@ export const buildFilesToCopy = [
   },
 ];
 
+/**
+ * Files to remove after the Build process completes.
+ */
 export const buildFilesToRemove = [
   "dist",
   "index.d.ts",
@@ -97,6 +104,9 @@ export const buildFilesToRemove = [
   "/src/server/types.js",
 ];
 
+/**
+ * Dependencies to remove from the application after compilation is completed.
+ */
 export const dependenciesToCleanup = [
   "@types/cors",
   "@types/express",