+++ CHANGELOG.md
@@ -1,9 +1,27 @@
-# Changelog
+\# Changelog
 
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [0.9.0] - 2020-06-11
+
+### 📦 Fixed Up Dependencies
+
+### Added
+
+- Comments in template files - [#22](https://github.com/alexlee-dev/create-mern-application/issues/22)
+- README `Built With` section - [#9](https://github.com/alexlee-dev/create-mern-application/issues/9)
+- GIF - [#33](https://github.com/alexlee-dev/create-mern-application/issues/33)
+
+### Changed
+
+- Revamped Documentation Website - [#34](https://github.com/alexlee-dev/create-mern-application/issues/34)
+
+### Removed
+
+### Fixed
 
 ## [0.8.0] - 2020-06-11